requirements.txt
streamlit
torch
torchvision
matplotlib
Pillow
numpy
scikit-learn
seaborn
opencv-python
